requirements.txt
numpy
Pillow
tabulate
tqdm
torch>=1.2
torchvision
